fix exports, add custum error
