--- controllers/contacts/deleteContact.js
@@ -1,27 +1,0 @@
-const contactsOperation = require('../../models/contacts');
-
-const deleteContact = async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const contact = await contactsOperation.removeContact(contactId);
-
-    if (!contact) {
-      const error = new Error('Not found');
-      error.status = 404;
-      throw error;
-    }
-
-    res.json({
-      status: 'success',
-      code: 200,
-      message: `contact deleted`,
-      data: {
-        result: contact,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-module.exports = deleteContact;
