--- controllers/contacts/getContacts.js
@@ -1,18 +1,0 @@
-const contactsOperation = require('../../models/contacts');
-
-const getContacts = async (req, res, next) => {
-  try {
-    const contacts = await contactsOperation.listContacts();
-    res.json({
-      status: 'success',
-      code: 200,
-      data: {
-        result: contacts,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-module.exports = getContacts;
