--- controllers/index.js
@@ -1,3 +1,0 @@
-const contacts = require('./contacts');
-
-module.exports = contacts;
