+++ models/contacts.json
@@ -1,1 +1,1 @@
-[{"id":"1","name":"Allen Raymond","email":"nulla.ante@vestibul.co.uk","phone":"(992) 914-3792"},{"id":"2","name":"Chaim Lewis","email":"dui.in@egetlacus.ca","phone":"(294) 840-6685"},{"id":"4","name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":"5","name":"Cyrus Jackson","email":"nibh@semsempererat.com","phone":"(501) 472-5218"},{"id":"6","name":"Abbot Franks","email":"scelerisque@magnis.org","phone":"(186) 568-3720"},{"id":"7","name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":"8","name":"Simon Morton","email":"dui.Fusce.diam@Donec.com","phone":"(233) 738-2360"},{"id":"9","name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":"10","name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"id":"2adb5223-d1bc-41d1-b886-0901f271a2ce","name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"id":"70104fc5-2be8-4d26-96aa-ab2ed68f689e","name":"Yurii Panasiuk","email":"yuriipanasiuk@gmail.com","phone":"(748) 206-2688"},{"id":"a6cf400c-3784-483d-9fb4-e0655613ca44","name":"Olia Panasiuk","email":"yuriipanasiuk@gmail.com","phone":"(748) 206-2688"},{"id":"a220ff51-537c-406a-9040-62dab607a756","name":"New User","email":"yuriipanasiuk@gmail.com","phone":"0238082034802384"},{"id":"1ea604de-3a23-49d8-9f0d-a1d84089c2a2","name":"New2 User","email":"yuriipanasiuk@gmail.com"},{"id":"5199f0b5-e133-40e0-b4e0-2ff4c2426ec6","name":"start node js adasdasd","email":"go-it@gmail.com","phone":"234-234243-234234"},{"id":"b52c8a97-161a-438f-80b3-1fe69e5e39ff","name":"Artem Panasiuk","email":"artem_panasiuk@gmail.com","phone":"234-234243-234234"},{"id":"f9303785-ed76-49be-bb55-1e8e9a0a16be","name":"Yurii Pansiuk","email":"yuriipanasiuk@gmail.com","phone":"096-800-39-01"}]
+[{"id":"1","name":"Allen Raymond","email":"nulla.ante@vestibul.co.uk","phone":"(992) 914-3792"},{"id":"2","name":"Chaim Lewis","email":"dui.in@egetlacus.ca","phone":"(294) 840-6685"},{"id":"4","name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":"5","name":"Cyrus Jackson","email":"nibh@semsempererat.com","phone":"(501) 472-5218"},{"id":"6","name":"Abbot Franks","email":"scelerisque@magnis.org","phone":"(186) 568-3720"},{"id":"7","name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":"8","name":"Simon Morton","email":"dui.Fusce.diam@Donec.com","phone":"(233) 738-2360"},{"id":"9","name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":"10","name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"id":"2adb5223-d1bc-41d1-b886-0901f271a2ce","name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"id":"70104fc5-2be8-4d26-96aa-ab2ed68f689e","name":"Yurii Panasiuk","email":"yuriipanasiuk@gmail.com","phone":"(748) 206-2688"},{"id":"a6cf400c-3784-483d-9fb4-e0655613ca44","name":"Olia Panasiuk","email":"olia_panasiuk@gmail.com","phone":"234-234243-234234"},{"id":"436bbb94-67ac-48ed-b9a1-15990fab3d5d","name":"go it academy","email":"go-it@gmail.com","phone":"123-123-123"}]
