+++ myErrors.js
@@ -1,0 +1,8 @@
+class MyError extends Error {
+  constructor(message, status) {
+    super(message);
+    this.status = status;
+  }
+}
+
+module.exports = MyError;
