+++ routes/api/contacts.js
@@ -1,15 +1,14 @@
 const express = require('express');
 const router = express.Router();
+const controllers = require('../../controllers/contactsControllers');
 
-const controllers = require('../../controllers');
-
-const { getContacts, getContactById, addNewContacts, deleteContact, updateContact } = controllers;
+const { getContacts, getContactById, addNewContact, deleteContact, updateContact } = controllers;
 
 router.get('/', getContacts);
 
 router.get('/:contactId', getContactById);
 
-router.post('/', addNewContacts);
+router.post('/', addNewContact);
 
 router.delete('/:contactId', deleteContact);
